refactor: Substitute DH Table in transforms before composing
It takes considerable calculation time (in the order of several seconds) to symbolically simplify the product of six 4x4 transform matrices each consisting of trigonometric expressions. By substituting some of the values for constants and angles in these trigonometric expressions for each individual matrix, the overall composite matrix is simplified much faster (down to the order of milliseconds).